implementation of search bar feature
the hero has a search bar that will search the fake store api for typed words.
the browse catalog button will scroll to the bottom of the page.

SUCCESS
The api search works locally
The html page renders well without issues in the console.

TODO
Still to render typed results to the page.